// src/types/contracts/RegistrationSimple.ts
export declare namespace RegistrationSimple {
  export type PassportStruct = {
    dgCommit: BigNumberish;
    dg1Hash: BytesLike;
    publicKey: BytesLike;
    passportHash: BytesLike;
    verifier: AddressLike;
  };

  export type PassportStructOutput = [
    dgCommit: bigint,
    dg1Hash: string,
    publicKey: string,
    passportHash: string,
    verifier: string
  ] & {
    dgCommit: bigint;
    dg1Hash: string;
    publicKey: string;
    passportHash: string;
    verifier: string;
  };
}