// src/types/contracts/StateKeeper.ts
export declare namespace StateKeeper {
  export type CertificateInfoStruct = { expirationTimestamp: BigNumberish };

  export type CertificateInfoStructOutput = [expirationTimestamp: bigint] & {
    expirationTimestamp: bigint;
  };

  export type PassportInfoStruct = {
    activeIdentity: BytesLike;
    identityReissueCounter: BigNumberish;
  };

  export type PassportInfoStructOutput = [
    activeIdentity: string,
    identityReissueCounter: bigint
  ] & { activeIdentity: string; identityReissueCounter: bigint };

  export type IdentityInfoStruct = {
    activePassport: BytesLike;
    issueTimestamp: BigNumberish;
  };

  export type IdentityInfoStructOutput = [
    activePassport: string,
    issueTimestamp: bigint
  ] & { activePassport: string; issueTimestamp: bigint };
}